Finished Car, tweaked Position.java a tiny bit

+++ Position.java
@@ -2,7 +2,7 @@ public class Position {
     private double x;
     private double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
